% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{assign_taxonomy_sintax}
\alias{assign_taxonomy_sintax}
\title{Run taxonomy assignment (SINTAX algorithm)}
\usage{
assign_taxonomy_sintax(
  seq_file,
  utax_db,
  confidence_threshold = 0.8,
  tmp_prefix = NULL,
  threads = 1
)
}
\description{
This requires \href{https://github.com/torognes/vsearch}{VSEARCH} to be installed.
}
