% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{check_system_requirements}
\alias{check_system_requirements}
\title{Check external software requirements and report missing programs}
\usage{
check_system_requirements()
}
\description{
Check external software requirements and report missing programs
}
