% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{create_report}
\alias{create_report}
\title{Create an Excel report}
\usage{
create_report(
  seq_tab,
  outfile,
  low_abund_threshold = 20,
  min_seqs_unknown = 10,
  n_curate = 4,
  n_show_depth = 6
)
}
\arguments{
\item{seq_tab}{data frame returned by \link{do_assign_taxonomy}}

\item{outfile}{the Excel output file}

\item{low_abund_threshold}{add 'low-coverage' to the issues list for samples
with less reads}

\item{n_curate}{max. number of sequences (polymorphism/haplotypes) from the
top taxon to list in the \emph{curation} section}

\item{n_show_depth}{max. number of sequences (polymorphism/haplotypes) from the
top taxon for which to output read numbers}

\item{min_seqs}{ignore samples found in reads (index combination) but not in the
samples table if they have less than the given number of reads}
}
\description{
Create an Excel report
}
