% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{dada2_denoise}
\alias{dada2_denoise}
\title{Run DADA2 denoising}
\usage{
dada2_denoise(
  x,
  dada_err,
  cores = 1,
  max_members = 1e+06,
  singleton_threshold = 5,
  ...
)
}
\arguments{
\item{x}{FASTQ file path or \emph{derep-class} object}

\item{dada_err}{error information returned by \link{dada_learn_errors}}

\item{max_members}{do random subsampling of overly large ASVs to
the specified number of sequences}

\item{singleton_threshold}{by default, singletons cannot form a new cluster,
but for very low-depth samples (< N duplicates of any sequence),
DETECT_SINGLETONS is turned on (see \link[dada2:setDadaOpt]{dada2::setDadaOpt}).
This increases sensitivity with InDel-rich Nanopore data and prevents that
a large proportion of sequences are discarded.}

\item{...}{Arguments passed to \link[dada2:dada]{dada2::dada}}
}
\value{
a data frame with the following columns:
\itemize{
\item \emph{sequence}
\item \emph{top_uniques}: abundance of top unique sequences
}
}
\description{
Runs DADA2 on a FASTQ file, removes chimeras and and returns
an abundance-sorted data frame of ASVs with some additional information
}
