% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{dada_learn_errors}
\alias{dada_learn_errors}
\title{Learn error rates}
\usage{
dada_learn_errors(fq_paths, omega_a = 1e-20, cores = 1, ...)
}
\description{
Calls \link[dada2:learnErrors]{dada2::learnErrors} with settings adjusted for Nanopore data and the clustering
procedure (\link{infer_barcode})
}
