% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{do_combine_alignments}
\alias{do_combine_alignments}
\title{Combine all BAM alignment files into a single one}
\usage{
do_combine_alignments(seq_tab, aln_dir, outdir, top_only = FALSE)
}
\arguments{
\item{seq_tab}{Sequence table as returned by \link{do_infer_all_barcodes} or downstream functions}

\item{aln_dir}{Alignment directory (as provided to \link{do_infer_all_barcodes})}

\item{out_prefix}{Output prefix for '.bam', '.bam.bai' and '.fasta' files}
}
\value{
Named list of output files
}
\description{
Combine all BAM alignment files into a single one
}
