% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{do_demux}
\alias{do_demux}
\title{Group sequences by sample index combination}
\usage{
do_demux(
  primer_search_fq,
  out_dir,
  sample_tab,
  error_threshold = 2.5,
  min_barcode_length = 50
)
}
\arguments{
\item{primer_search_fq}{One or multiple paths from \link{do_primer_search}
(\code{trimmed_fq} entry in the returned list).}
}
\value{
\code{sample_tab} with a \code{reads_path} column and additional rows
for samples found in the reads but not in the sample table
}
\description{
Group sequences by sample index combination
}
\details{
See \link{do_trim_demux} for details on the other arguments

The read files in \code{out_dir} are named as follows: \code{fprimer-fidx--rprimer-ridx.fastq.gz}

Requires \href{https://github.com/markschl/seqtool}{seqtool} v0.4 or higher
}
