% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{do_infer_all_barcodes}
\alias{do_infer_all_barcodes}
\title{Infer barcode sequences for all samples}
\usage{
do_infer_all_barcodes(
  seq_tab,
  dada_err,
  aln_out = NULL,
  tmp_dir = NULL,
  parallel_lapply_fn = NULL,
  ...,
  cores = 1
)
}
\arguments{
\item{seq_tab}{Sequence metadata table returned by \link{do_trim_demux} or \link{do_demux}}

\item{dada_err}{Result of \link{dada_learn_errors}}

\item{aln_out}{Optional output directory for BAM alignments (none saved if \code{NULL})}

\item{tmp_dir}{Optional temporary directory (e.g. \href{https://en.wikipedia.org/wiki/RAM_drive}{RAM drive}
such as \verb{/run/user/1000/DadaNanoBC} on Linux, for faster processing)}
}
\value{
Input table (\code{seq_tab}) with a new column \code{clustering}, which is a list
of data frames as returned by \link{infer_barcodes}
}
\description{
High-level function calling \link{infer_barcodes} on every sample,
Parallel processing possible with \code{cores} > 1. Custom parallel processing
frameworks can be plugged in by providing \code{parallel_lapply_fn}.
}
