% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{do_assign_compare_taxonomy}
\alias{do_assign_compare_taxonomy}
\alias{do_assign_taxonomy}
\alias{do_compare_taxonomy}
\title{Run the taxonomy assignment and compare with provided taxonomic names of specimens}
\usage{
do_assign_compare_taxonomy(
  seq_tab,
  db_file,
  gbif_cache_file,
  tmp_dir = NULL,
  confidence_threshold = 0.8,
  summary_ranks = NULL,
  known_contaminants = NULL,
  likely_kingdom = NULL,
  contam_rank_delta = 3,
  cores = 1
)

do_assign_taxonomy(
  seq_tab,
  db_file,
  tmp_dir = NULL,
  confidence_threshold = 0.8,
  summary_ranks = NULL,
  cores = 1
)

do_compare_taxonomy(
  seq_tab,
  seq_lineages,
  gbif_cache_file,
  known_contaminants = NULL,
  likely_kingdom = NULL,
  contam_rank_delta = 3
)
}
\arguments{
\item{seq_tab}{Sequence table as returned by \link{do_infer_all_barcodes} or downstream functions}

\item{db_file}{Sequence database with UTAX-formatted headers (optionally compressed);
see \link{load_taxdb}}

\item{gbif_cache_file}{Path to cache file for saving GBIF taxonomy lookup results
(may be shared across datasets) (see \link{get_gbif_taxa})}

\item{tmp_dir}{optional temporary directory for saving intermediate sequence files}

\item{confidence_threshold}{SINTAX bootstrap threshold (default: 0.8 / 80\%)}

\item{summary_ranks}{optional named character vector with ranks that should
show up in the lineages (default: determine automatically)}

\item{known_contaminants}{nested list specifying known contaminant names,
such as \code{list(family=c('Aspergillaceae', ...), genus=c(...))}}

\item{likely_kingdom}{kingdom name providing some guidance for the GBIF
name search (see \link{get_gbif_taxa})}

\item{contam_rank_delta}{Require at least N additional ranks to be matching
between the provided (e.g. morphology-based) and the auto-assigned
sequence-based taxonomic lineages for a taxon to be determined as the
"correct" taxon in the mix, and the top taxon being down-ranked.
This setting affects the "sensitivity" of the contaminant detection
(lower: more sensitive, but also more false classifications,
higher: some contamination may not be recognized).
The default of 3 seems to work well in most cases.}
}
\value{
\code{do_assign_compare_taxonomy} returns
\code{seq_tab}, with the nested data frames in \code{clustering} \strong{reordered} as such
that dominant taxa flagged as contamination are moved to the bottom.
The following new columns are added:
\itemize{
\item \emph{unique_id}: unique sequence id in the form: \code{u1}, \code{u2}, etc.
\item \emph{taxon}: auto-inferred taxon name
(\emph{may not be correct, confirm with BLAST or other identification methods})
\item \emph{short_lineage}: taxonomic lineage of the auto-inferred taxon
(provided or auto-inferred \code{summary_ranks} shown)
\item \emph{is_contaminant}: (logical) \code{TRUE} for taxa recognized as contamination
\item \emph{matching_ranks}: Number of matching taxonomic ranks in the comparison
between the provided and auto-assigned sequence-based taxonomy
(usually: kingdom/domain, phylum, class, order, family, genus, species).
The comparison can only be done up to the highest defined rank in either
lineage.
\item \emph{mismatching_ranks}: Number of non-matching ranks in the comparison
between the provided and auto-assigned sequence-based taxonomy
\item \emph{unspecific}: \code{TRUE} for all taxa that are not at the top in the clusters table
after the contamination ranking
}

The main \code{seq_tab} table further contains some columns with data from the
top sequence: \emph{top_seq_matching_ranks}, \emph{top_seq_mismatching_ranks},
\emph{top_seq_taxon}, \emph{top_seq_short_lineage}
}
\description{
This function runs \code{do_assign_taxonomy} and \code{do_compare_taxonomy} sequentially
(functions not documented).
}
\details{
Sequences are flagged as contamination (\emph{is_contaminant}) if:
\itemize{
\item there exists a less abundant taxon with the correct kingdom
(compared to \emph{taxon}, GBIF name comparison guided by \code{likely_kingdom})
\item a taxon has has at least \code{contam_rank_delta} more consistent (matching)
taxonomic ranks in the GBIF lineage than the top taxon
(explanation for \emph{matching_ranks} above); undefined ranks are excluded
from the comparison
}
}
