% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{infer_barcode}
\alias{infer_barcode}
\title{Infer barcode sequence(s) from raw reads in a single sample}
\usage{
infer_barcode(
  fq,
  dada_err,
  alignment_prefix,
  id_prefix = NULL,
  tmp_dir = NULL,
  dada_omega_a = c(1e-20, 1e-10, 0.01),
  omegaA_iter_threshold = 1000,
  dada_min_identical = 2,
  dada_min_n0 = 4,
  min_seq_abund = 3,
  max_sample_depth = 5000,
  consensus_max_depth = 3000,
  consensus_threshold = 0.65,
  consensus_by_qual = TRUE,
  homopoly_fix_min_ident = 4,
  min_homopoly_len = 6,
  fixed_cluster_threshold = 0.97,
  taxa_cluster_threshold = fixed_cluster_threshold,
  cluster_single_linkage = TRUE,
  min_variant_freq = 0.2,
  split_min_identical = 4,
  max_split_ratio = 3,
  cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{fq}{demultiplexed FASTQ file}

\item{dada_err}{output from \link{dada_learn_errors}}

\item{alignment_prefix}{Optional output prefix for BAM alignment files and
FASTA references for manual inspection and/or downstream analyses}

\item{id_prefix}{Name prefix for sequence IDs in the consensus FASTA and BAM alignment files}

\item{tmp_dir}{Optional path to a temporary directory (details in \link{set_global_opts})}

\item{dada_omega_a}{OMEGA_A parameter value(s) to use for the denoising
(see \link[dada2:setDadaOpt]{dada2::setDadaOpt}):
Can be a vector of increasing values (the default), which are tried
sequentially until a grouping with well-separated haplotypes and
non-ambiguous consensus sequence(s) is found for the \emph{top taxon}
(\emph{note}: others currently ignored)}

\item{omegaA_iter_threshold}{Only evaluate multiple \code{dada_omega_a} values if the
read depth is not larger than the given threshold. For large samples, it is
better to avoid multiple denoising rounds, and it is anyway likely that DADA2
will already find all sequence variants in the first round (with lowest OMEGA_A).
Haplotype variation with InDels may still be found with 'try_split_haplotypes',
but not with DADA2 denoising at high sensitivity.}

\item{dada_min_identical}{Minimum number of identical sequences required to do a
DADA2 denoising. Below this threshold, switch to simple fixed-threshold
clustering instead (at \code{fixed_cluster_threshold}) and report the consensus
without further attempting any haplotype splitting.}

\item{dada_min_n0}{Minimum number of error-free reads needed to retain the DADA2 clustering
(\code{n0} in dada-class $clustering information).
Switch to fixed-threshold clustering + consensus method below this threshold
(see \code{fixed_cluster_threshold}).
\emph{Note}: \emph{error-free} does not mean \emph{identical}, as DADA2 only considers
substitutions, and there can still be InDels, so \code{n0} is always equal or higher
to the number of identical sequences.}

\item{min_seq_abund}{minimum number of supporting reads needed for any barcode
sequence variant to be included in the results}

\item{max_sample_depth}{read a maximum of \code{max_sample_depth} demultiplexed reads
from the input file (\code{fq}) for the denoising/clustering}

\item{consensus_max_depth}{maximum number sequences mapped against the
clusters to infer the consensus sequence (if there are more, a random sample
is taken)}

\item{consensus_threshold}{require at least the given proportion of bases
to be identical at every alignment column for an unambiguous consensus call
(values below the default 60\% might be problematic)}

\item{consensus_by_qual}{Consider the read quality scores when building the
consensus with \href{https://www.htslib.org/doc/samtools-consensus.html}{samtools consensus}.
If \code{TRUE}, the relative base frequencies are weighted by the Phred quality scores.}

\item{homopoly_fix_min_ident}{Minimum number of identical sequences required to
attempt adjusting ambiguous homopolymer sequences in the consensus.
The homopolymer run length of the most frequent sequence is chosen.}

\item{min_homopoly_len}{Minimum length a homopolymer region needs to have in order
to attempt "fixing" an ambiguous consensus in that region
(see also \code{homopoly_fix_min_ident})}

\item{fixed_cluster_threshold}{Similarity threshold for grouping sequence variants
per taxon. Single-linkage clustering is applied, with the default threshold of
0.97, the maximum divergence between any two sequences can be 3\%.}

\item{taxa_cluster_threshold}{Similarity threshold
and for post-clustering of DADA2 ASVs to group potential haplotypes of the
same sequenced organism together.}

\item{cluster_single_linkage}{Whether to apply single-linkage fixed-threshold
clustering for low-depth samples.
Clusters grow as long as any two sequences have at least \code{fixed_cluster_threshold}
similarity. If \code{FALSE}, the all cluster members are compared to one centroid
sequence and remain more concise (=complete-linkage clustering).
On (\code{TRUE}) by default, as this appears to work well for low-coverage samples.}

\item{min_variant_freq}{frequency threshold to consider separate sequences (ASVs)
of the same taxon (as clustered with \code{taxa_cluster_threshold}) as "real" polymorphisms
(haplotypes), not noise.}

\item{split_min_identical}{min. number of identical sequences that needs to
support \emph{both} of top two unique sequences in a sample in order to attempt
haplotype splitting (with these sequences as new references).}

\item{max_split_ratio}{only split an ASV into 2 parental haplotypes if the ratio
of larger:smaller is up to \code{max_split_ratio}.
While the haplotypes might not always have an exact 1:1 ratio, this
constraint still enforces a certain balancing of haplotype abundances.
More unbalanced ratios are only possible if DADA2 is rerun with higher sensitivity
(see \code{dada_omega_a}).}
}
\value{
Returns a data frame with at least the following columns:
\itemize{
\item \emph{id}: Descriptive sequence ID, e.g.: 'taxon1_seq2'
\item \emph{full_id}: The ID used in the BAM and FASTA output files: \emph{id} prefixed with \code{id_prefix}
\item \emph{taxon_num}: Nth (putative) taxon (integer; see also \code{taxa_cluster_threshold})
\item \emph{sequence}: DADA2 ASV or most abundant unique split sequence
(\code{NA} in case of fixed-threshold clustering)
\item \emph{consensus}: consensus sequence of haplotype member sequences
\item \emph{consensus_ambigs}: number of ambiguous bases in the consensus (see \code{consensus_threshold})
\item \emph{consensus_diffs}: edit distance between consensus and the representative sequence
(DADA2 ASV) or dominant sequence (after haplotype splitting/fixed-threshold clustering)
\item \emph{homopolymer_adjustments}: number of homopolymer locations that were "fixed"
by inferring the repeat number from the most abundant sequence instead of relying
on the consensus (containing Ns)
\item \emph{abundance}: number of reads clustering with the given sequence
(up to \code{max_sample_depth})
\item \emph{n_mapped}: number of reads mapped to the barcode sequences
(usually the same or very similar to \emph{abundance})
\item \emph{n0} from DADA2: number of error-free reads (no substitutions, there may still be InDels)
\item \emph{max_identical}: max. number of identical reads
\item \emph{method}: method from which the haplotype emerged:
one of 'dada', 'dada_split' (if haplotype splitting was done), 'fixed_cluster'
\item \emph{is_rare}: \code{TRUE} if the relative frequency of a sequence within the same
putative taxon is below \code{min_variant_freq}
}
}
\description{
Applies a clustering workflow to the provided raw reads, based on
DADA2 denoising and/or fixed-threshold clustering to infer the correct
barcode sequence(s). Further groups the resulting sequences by putative
taxon.
}
