% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{init_pipeline}
\alias{init_pipeline}
\title{Scaffold the Targets Pipeline}
\usage{
init_pipeline(path = ".", analysis_dir = "analysis", workers = NULL)
}
\arguments{
\item{path}{Where to create the \verb{_targets.R} and other files}

\item{analysis_dir}{Analysis directory}
}
\description{
Scaffold the Targets Pipeline
}
