% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{load_taxdb}
\alias{load_taxdb}
\title{Download a taxonomic database and save it in UTAX format}
\usage{
load_taxdb(
  db_url,
  db_type = c("unite", "qiime_fasta", "qiime_qza"),
  db_file,
  db_tax_url = NULL,
  ...
)
}
\arguments{
\item{db_url}{Character vector with one or more file URLs; sequences from all
URLs are combined into one final database.}

\item{db_type}{Database type (character) (see details)}

\item{db_file}{Output file (will be GZIPped FASTA, \code{.fasta.gz})}

\item{db_tax_url}{(optional) URL(s) pointing to file(s) containing the taxonomy for the
sequences in the files at \code{db_url} (currently implemented for \emph{qiime_qza} format)}
}
\description{
Download a taxonomic database and save it in UTAX format
}
\details{
Currently, three database types are implemented (more may follow):
\itemize{
\item \emph{"unite"}: UNITE ITS database (https://unite.ut.ee/repository.php):
choose a dataset (e.g. \href{https://doi.plutof.ut.ee/doi/10.15156/BIO/3301243}{all Eukaryotes}),
click on the file at \emph{Downloads} and copy the actual file URL once the file
starts downloading in the browser.
The SH sequence set with a \emph{dynamic clustering threshold} is chosen from the archive
and imported.
\item \emph{"qiime_fasta"}: FASTA file with lineages in headers in the form
\verb{<seqid> rank1__name; rank2__name; ...}
whereby \code{rank1} and \code{rank2} are one-letter codes such as \code{k} (for kingdom), \code{g} (for genus), etc.
If only sequences are in the file, specify the URL of a tab-delimited file
mapping sequence IDs to lineages.
One example of database providing such files \href{https://gtdb.ecogenomic.org}{GTDB}
(e.g. \href{https://data.gtdb.aau.ecogenomic.org/releases/release226/226.0/genomic_files_reps}{here})
\item \emph{"qiime_qza"}: \href{https://amplicon-docs.qiime2.org/en/stable/explanations/archives.html}{QIIME2 QZA} format,
which is essentially a ZIP file containing a \emph{qiime_fasta}-formatted FASTA
file, while another QZA file should contain the tab-delimited taxonomy (provide with \code{db_tax_url}).
}

This function may be expanded and made more flexible in the future
}
