% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{parse_primer_tab}
\alias{parse_primer_tab}
\title{Parse a data frame with primer information}
\usage{
parse_primer_tab(primer_tab, amplicons = NULL)
}
\arguments{
\item{primer_tab}{Primer table (data frame-like; see details)}

\item{amplicons}{Character vector of amplicons pooled in the library;
their names should be in the form \code{forward-reverse} (primer names should not contain dashes)}
}
\value{
a named list of amplicon metadata (named by amplicons = \code{forward-reverse}).
Each amplicon-specific item is another list with \emph{two entries} (for forward and reverse),
which are again lists with following entries:
\itemize{
\item \emph{primer}: named vector of primer sequences
\item \emph{index}: named vector of index sequences
\item \emph{index_len}: the length of the index sequences (no variable length currently allowed,
but lengths may vary between forward and reverse sample indexes)
}
}
\description{
Parse a data frame with primer information
}
\details{
Columns of \code{primer_tab}:
\itemize{
\item \emph{primer-index}: Primer name and sample index name, delimited by a dash;
the primer names \strong{must not contain dashes} and both \emph{primer} and \emph{index}
should correspond to the names in the \emph{amplicon} column of the sample table
(see \link{parse_sample_tab}).
\item \emph{index_seq} sample index sequence
\item \emph{primer_seq} primer sequence (repeated)
}

If \code{amplicon} is not provided, the forward/reverse primers are assumed to be
in order. \code{amplicon} may be obtained from the \emph{amplicon} column of the sample
table (see \link{parse_sample_tab}).
}
