% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{parse_sample_tab}
\alias{parse_sample_tab}
\title{Parse a data frame with sample metadata}
\usage{
parse_sample_tab(sample_tab)
}
\arguments{
\item{sample_tab}{The sample table (data frame-like, see details)}
}
\value{
a valid sample metadata table (data frame)
with sample names de-duplicated (if necessary)
}
\description{
Parse a data frame with sample metadata
}
\details{
Required columns:
\itemize{
\item Plate
\item \emph{plate}: A01-H12)
\item \emph{amplicon}: \code{forward-reverse}, name as in 'primer_index' column of primer table;
see \link{parse_primer_tab}
\item \emph{indexes}: \code{forward-reverse}, name as in 'primer_index' column of primer table;
see \link{parse_primer_tab}
\item \emph{sample}: sample name, should not be duplicated
\item \emph{sample_type}: Sample type such as 'negative control', will show up in reports;
can be left empty
}

Optional columns:
\itemize{
\item \emph{taxon}: Known or suspected taxon (any rank), will be compared to
the sequence-based identification and used for detecting contamination
\item \emph{known sequence}: Already known sequence (if any), e.g. from a previous Sanger
sequencing; will be compared with the Nanopore-derived sequence
}

\subsection{Note on amplicon multiplexing}{

Primers are searched in the order that amplicons appear in the sample table
With \emph{nested amplicons}, the shorter one should be placed at the end.
Also place amplicons with \emph{little data} before other amplicons with similar
primers to make sure that they don't get "swallowed" in case of unspecific
primer or sample index matching.
}
}
