% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{read_xlsx_primer_tab}
\alias{read_xlsx_primer_tab}
\title{Read primer information from Excel file}
\usage{
read_xlsx_primer_tab(meta_file, sheet_name, ...)
}
\description{
Reads primer information from an Excel metadata file and passes it on to \link{parse_primer_tab}
}
