% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{read_xlsx_sample_tab}
\alias{read_xlsx_sample_tab}
\title{Read sample metadata from Excel file}
\usage{
read_xlsx_sample_tab(meta_file, sheet_name)
}
\description{
Reads a sample table from an Excel metadata file and passes it on to \link{parse_sample_tab}
}
