% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{set_global_opts}
\alias{set_global_opts}
\title{Set global options}
\usage{
set_global_opts(...)
}
\description{
Convenience function for setting global options such as program paths.
}
\details{
The following options can be set:
\itemize{
\item Required program paths (\code{samtools}, \code{minimap2}, \code{seqtool}, \code{vsearch})
\item \code{tmp_dir}: Path to a custom temporary directory for short-lived
data and not overly large data. On Linux, the default is to place these files
in a subdirectory of \verb{$XDG_RUNTIME_DIR} (usually \verb{/run/user/<userid>}, tmpfs).
}

Under the hood, this simply does \code{options(DadaNanoBC.opt = 'value', ...)}.

The same options can be set as environment variables:
\code{DadaNanoBC_opt=value} (in R: \code{Sys.setenv(DadaNanoBC_opt = 'value')}.
\code{set_global_opt} always has priority over environment variables.
}
\examples{
set_global_opts(samtools = '/path/to/samtools', tmp_dir = '/tmp/nanopore-barcoding')

}
