% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{set_program_path}
\alias{set_program_path}
\title{Set the path to one or more programs}
\usage{
set_program_path(...)
}
\description{
Set the path to one or more programs
}
\examples{
set_program_path(samtools = '/path/to/samtools', minimap2 = '/path/to/minimap2')

}
