% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{init_pipeline}
\alias{init_pipeline}
\title{Scaffold the Targets Pipeline}
\usage{
init_pipeline(path = ".", bash = FALSE, analysis_dir = "analysis")
}
\arguments{
\item{path}{Where to create the \verb{_targets.R} and other files}

\item{bash}{(logical) if \code{TRUE}, initialize in "Bash mode", which means
that an \code{infer_barcodes} Bash script is copied along with \verb{_targets.R}}

\item{analysis_dir}{Analysis directory}
}
\description{
Copies \verb{_targets.R} to current directory (or \code{path}) and initializes
an analysis directory (without overwriting files).
}
